WIP Adding Dual extruder compatibility
To become compatible with dual extrusion, the script must be able to
export individual model files specific to the different extruders. To do
this, we add an option to only render a subset of the model.

This WIP updates the params form to specify which section to render,
adds 3 rendering strategies, and begins to parameterize the model to use
for Dots.

The Dots model must be parameterized to support dualstrusion with the
CreeperHead model.